# feat: improve performance on linkRecursiveCode