fix #331

before:
```
18:47:08 | 200 |  6.702s |       127.0.0.1 | GET     | /api/nonuser/all/products
18:47:09 | 200 |     1ms |       127.0.0.1 | GET     | /api/nonuser/all/products
```

after:
```
00:58:50 | 200 |   137ms |       127.0.0.1 | GET     | /api/nonuser/all/products
00:59:00 | 200 |     1ms |       127.0.0.1 | GET     | /api/nonuser/all/products
```